Fixed the code problems
Added a debug mode that toggles off the spamming of data so that debug
messages can be seen. Also fixed a small error that was causing messages
to not go through in the parseSerial() function. Also changed the
delimiter in the messages to the '&' character to solve an odd newline
character problems different systems were having communicating to the
Arduino.